// src/services/types/useCase/bug-tracking.ts
export type BugTrackingApiResponse = {
  _id: string
  bug_tracking: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
    }
    section2: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      image1: string
      title2_en: string
      title2_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
    }
    section3: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      text4_en: string
      text4_fa: string
      text5_en: string
      text5_fa: string
      text6_en: string
      text6_fa: string
      text7_en: string
      text7_fa: string
      image2: string
    }
    section4: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      text4_en: string
      text4_fa: string
      text5_en: string
      text5_fa: string
      text6_en: string
      text6_fa: string
      text7_en: string
      text7_fa: string
      text8_en: string
      text8_fa: string
      image2: string
    }
    section5: {
      title1_en: string
      title1_fa: string
      image1: string
      image2: string
      image3: string
      title2_en: string
      title2_fa: string
      title3_en: string
      title3_fa: string
      title4_en: string
      title4_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
    }
    section6: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      text4_en: string
      text4_fa: string
      text5_en: string
      text5_fa: string
      text6_en: string
      text6_fa: string
      text7_en: string
      text7_fa: string
      text8_en: string
      text8_fa: string
      image2: string
    }
    section7: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      title2_en: string
      title2_fa: string
      text2_en: string
      text2_fa: string
      title3_en: string
      title3_fa: string
      text3_en: string
      text3_fa: string
      title4_en: string
      title4_fa: string
      text4_en: string
      text4_fa: string
      title5_en: string
      title5_fa: string
      text5_en: string
      text5_fa: string
      title6_en: string
      title6_fa: string
      text6_en: string
      text6_fa: string
    }
    section8: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
  __v: number
}